Add majority of circuit required to implement hiccup mode current limiting.
Hiccup mode current limiting works by detecting an over-current condition and
then disabling the output. The output remains disabled for a set period of
time after which it is turned back on. If the fault which caused too much
current to be drawn is still present then the process repeats. The output will
effectively 'hiccup' until the fault is removed.

The added components basically implement a 555 timer monostable with 3
comparators. The fourth comparator determines if the voltage across the sense
resistor is high enough to start the current limiting. The voltage from the
sense resistor passes through a low-pass filter to remove short spikes of
current while will be present when using the H-bridge for PWM.

The far-right comparator is just an inverter with hysteresis to drive the gate
of a p-channel MOSFET (still to be added). The comparator to the left of it is
configured as an RS flip-flop, which is set/reset by pulling its inputs low.
One of the left-hand comparators acts as a timer, measuring the voltage across
a capacitor and comparing it to a fixed reference. The other comparator is
what senses the over-current.