Add values to termination resistors. Insert more term. resistors on 'in' line
As the 'In' line needs to cope with 12V being injected into it (if the
motor board is plugged into a powered SRIC socket) the 100 ohm
termination resistor needs to be able to handle a fairly high power.
Around 1.5W at 12V and in the worst case 2.56W at 16V. The four 100
ohm 0.5W resistors can therefore cope with 2W continuous.