Swap ITL and IBL signals from MSP to H-Bridge
The MSP can only generate PWM on the 'TB1' and 'TB2' outputs. It
uses the compare unit associated with 'TB0' to set the upper limit
of the counter, and also to control toggling the other two outputs.

It is now only possible to modulate the upper FETs, the lower ones
are just switched on permanently.