Add locating tabs and slots to lower half of sides and base
The slots and tabs are all different sizes due to the non-uniform
positioning of the T-slots and corresponding holes in the base.
By being different sizes it means the side panels can not get mixed
up and it's only possible to assemble the lower half in two ways
(one upside down/inside out).

The long tabs provide extra strength to the side panels where there
would otherwise be a small strip of acrylic.