Break ground plane around current sense resistor
I found that the copper between the sense resistor and the reverse
polarity protection FET was dropping around 5mV. As the ground for
the comparator circuit was being taken from half way along that bit
of copper its ground potential ended up 3-4mV lower than that present
at the sense resistor when 5A was flowing into the load. This has the
effect of reducing the current limit from 5.5A to ~4.5A.

The ground path around the sense resistor has now been broken up and
the comparator ground is referenced with respect to the lower terminal
of the sense resistor. This means that its ground potential will be
raised slightly above the 'power' ground potential going into the
reverse polarity protection FET but this is not a problem and doesn't
adversely affect the circut in any way.