Increase FET driver decoupling caps from 470nF to 4.7uF
There is quite a bit of noise on the 12V rail when the output is shorted
and this gets fed into the FET driver ICs causes all manner of bad things
to happen. One of the ICs blew up when the output was shorted until the
decoupling caps were increased to this value.