Don't populate the feedback components
As we've never created a board to utilise the feedback connector we decided to
not populate the feedback components in the next batch. The case will be
modified to suit the lack of feedback connector.

Due to some funky goings on with the new units-in-pcbfile feature and gsch2pcb I
used the 'import schematics' feature of PCB to update the layout from the
schematic in this commit. Hence the addition of the two schematic src layout
level attributes and all the other attributes it added to the components
themselves.